add
add

+++ app/src/main/java/com/walle/gankio/utils/BitmapUtil.java
@@ -183,7 +183,7 @@ public class BitmapUtil {
        /* if(widthOrg > edgeLength && heightOrg > edgeLength)
         {*/
         //压缩到一个最小长度是edgeLength的bitmap
-        int longerEdge = (int)(edgeLength * Math.max(widthOrg, heightOrg) / Math.min(widthOrg, heightOrg));
+        int longerEdge = edgeLength * Math.max(widthOrg, heightOrg) / Math.min(widthOrg, heightOrg);
         int scaledWidth = widthOrg > heightOrg ? longerEdge : edgeLength;
         int scaledHeight = heightOrg;
         //widthOrg > heightOrg ? edgeLength : longerEdge;
